# swagger.yaml
swagger: '2.0'
info:
  description: 'Assisted installation'
  version: 1.0.0
  title: AssistedInstall
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: api.openshift.com
basePath: /api/assisted-install
...
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  agentAuth:
    type: apiKey
    in: header
    name: X-Secret-Key
  userAuth:
    type: apiKey
    in: header
    name: Authorization
  urlAuth:
    type: apiKey
    in: query
    name: api_key
  imageAuth:
    type: apiKey
    in: header
    name: Image-Token
security:
  - userAuth: [admin, user]
paths:
...
  /v1/clusters/{cluster_id}/events:
    get:
      tags:
        - events
      security:
        - userAuth: [admin, read-only-admin, user]
        - urlAuth: []
      description: Lists events for a cluster.
      operationId: ListEvents
      parameters:
        - in: path
          name: cluster_id
          description: The cluster to return events for.
          type: string
          format: uuid
          required: true
        - in: query
          name: host_id
          description: A host in the specified cluster to return events for.
          type: string
          format: uuid
          required: false
        - in: query
          name: categories
          description: A comma-separated list of event categories.
          type: array
          items:
            type: string
          required: false
      responses:
        "200":
          description: Success.
          schema:
            $ref: '#/definitions/event-list'
        "401":
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        "403":
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        "404":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "405":
          description: Method Not Allowed.
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Error.
          schema:
            $ref: '#/definitions/error'
  /v1/clusters/{cluster_id}/monitored_operators:
    get:
      tags:
        - operators
        - installer
      security:
        - agentAuth: []
      description: Lists operators to be monitored for a cluster.
      operationId: ListOfClusterOperators
      parameters:
        - in: path
          name: cluster_id
          description: The cluster to return operators for.
          type: string
          format: uuid
          required: true
        - in: query
          name: operator_name
          description: An operator in the specified cluster to return its data.
          type: string
          required: false
      responses:
        "200":
          description: Success.
          schema:
            $ref: '#/definitions/monitored-operators-list'
        "401":
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        "403":
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        "404":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "405":
          description: Method Not Allowed.
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Error.
          schema:
            $ref: '#/definitions/error'
    put:
      tags:
        - operators
        - installer
      security:
        - agentAuth: []
      description: Controller API to report of monitored operators.
      operationId: ReportMonitoredOperatorStatus
      parameters:
        - in: path
          name: cluster_id
          description: The cluster whose operators are being monitored.
          type: string
          format: uuid
          required: true
        - in: body
          name: report-params
          description: The operators monitor report.
          required: true
          schema:
            $ref: '#/definitions/operator-monitor-report'
      responses:
        "200":
          description: Success.
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        "403":
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        "404":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "405":
          description: Method Not Allowed.
          schema:
            $ref: '#/definitions/error'
        "409":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "503":
          description: Unavailable.
          schema:
            $ref: '#/definitions/error'
...
  /v1/supported-operators:
    get:
      tags:
        - operators
      operationId: ListSupportedOperators
      description: Retrieves the list of supported operators.
      responses:
        "200":
          description: Success.
          schema:
            type: array
            items:
              type: string
        "401":
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        "403":
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        "500":
          description: Error.
          schema:
            $ref: '#/definitions/error'
  /v1/supported-operators/{operator_name}:
    get:
      tags:
        - operators
      security:
        - userAuth: [admin, read-only-admin, user]
      description: Lists properties for an operator.
      operationId: ListOperatorProperties
      parameters:
        - in: path
          name: operator_name
          description: The operator name.
          type: string
          required: true
      responses:
        "200":
          description: Success.
          schema:
            $ref: '#/definitions/operator-properties'
        "401":
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        "403":
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        "404":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Error.
          schema:
            $ref: '#/definitions/error'
...
  /v2/clusters/{cluster_id}/monitored-operators:
    get:
      tags:
        - operators
        - installer
      security:
        - agentAuth: []
      description: Lists operators to be monitored for a cluster.
      operationId: V2ListOfClusterOperators
      parameters:
        - in: path
          name: cluster_id
          description: The cluster to return operators for.
          type: string
          format: uuid
          required: true
        - in: query
          name: operator_name
          description: An operator in the specified cluster to return its data.
          type: string
          required: false
      responses:
        "200":
          description: Success.
          schema:
            $ref: '#/definitions/monitored-operators-list'
        "401":
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        "403":
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        "404":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "405":
          description: Method Not Allowed.
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Error.
          schema:
            $ref: '#/definitions/error'


    put:
      tags:
        - operators
        - installer
      security:
        - agentAuth: []
      description: Controller API to report of monitored operators.
      operationId: v2ReportMonitoredOperatorStatus
      parameters:
        - in: path
          name: cluster_id
          description: The cluster whose operators are being monitored.
          type: string
          format: uuid
          required: true
        - in: body
          name: report-params
          description: The operators monitor report.
          required: true
          schema:
            $ref: '#/definitions/operator-monitor-report'
      responses:
        "200":
          description: Success.
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        "403":
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        "404":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "405":
          description: Method Not Allowed.
          schema:
            $ref: '#/definitions/error'
        "409":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "503":
          description: Unavailable.
          schema:
            $ref: '#/definitions/error'
...
  /v2/events:
    get:
      tags:
        - events
      security:
        - userAuth: [admin, read-only-admin, user]
        - urlAuth: []
      description: Lists events for a cluster.
      operationId: v2ListEvents
      parameters:
        - in: query
          name: cluster_id
          description: The cluster to return events for.
          type: string
          format: uuid
          required: False
        - in: query
          name: host_id
          description: A host in the specified cluster to return events for.
          type: string
          format: uuid
          required: false
        - in: query
          name: infra_env_id
          description: The infra-env to return events for.
          type: string
          format: uuid
          required: false
        - in: query
          name: categories
          description: A comma-separated list of event categories.
          type: array
          items:
            type: string
          required: false
      responses:
        "200":
          description: Success.
          schema:
            $ref: '#/definitions/event-list'
        "401":
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        "403":
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        "404":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "405":
          description: Method Not Allowed.
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Error.
          schema:
            $ref: '#/definitions/error'
...
definitions:
...
  monitored-operator:
    type: object
    properties:
      cluster_id:
        type: string
        format: uuid
        description: The cluster that this operator is associated with.
        x-go-custom-tag: gorm:"primaryKey"
      name:
        type: string
        description: Unique name of the operator.
        x-go-custom-tag: gorm:"primaryKey"
      namespace:
        type: string
        description: Namespace where to deploy an operator. Only some operators require a namespace.
      subscription_name:
        type: string
        description: The name of the subscription of the operator.
      operator_type:
        $ref: '#/definitions/operator-type'
      properties:
        type: string
        description: Blob of operator-dependent parameters that are required for installation.
        x-go-custom-tag: gorm:"type:text"
      timeout_seconds:
        type: integer
        description: Positive number represents a timeout in seconds for the operator to be available.
      status:
        $ref: '#/definitions/operator-status'
      status_info:
        type: string
        description: Detailed information about the operator state.
      status_updated_at:
        type: string
        format: date-time
        x-go-custom-tag: gorm:"type:timestamp with time zone"
        description: Time at which the operator was last updated.
  operator-monitor-report:
    type: object
    properties:
      name:
        type: string
        description: Unique name of the operator.
      status:
        $ref: '#/definitions/operator-status'
      status_info:
        type: string
        description: Detailed information about the operator state.
  operator-type:
    type: string
    enum: ['builtin', 'olm']
    description: Kind of operator. Different types are monitored by the service differently.
  operator-status:
    type: string
    enum: ['failed', 'progressing', 'available']
    description: Represents the operator state.
...
  monitored-operators-list:
    type: array
    items:
      $ref: '#/definitions/monitored-operator'
...
  event-list:
    type: array
    items:
      $ref: '#/definitions/event'
  event:
    type: object
    required:
      - severity
      - message
      - event_time
    properties:
      cluster_id:
        type: string
        format: uuid
        description: Unique identifier of the cluster this event relates to.
        x-go-custom-tag: gorm:"index"
        x-nullable: true
      host_id:
        type: string
        format: uuid
        description: Unique identifier of the host this event relates to.
        x-go-custom-tag: gorm:"index"
        x-nullable: true
      infra_env_id:
        type: string
        format: uuid
        description: Unique identifier of the infra-env this event relates to.
        x-go-custom-tag: gorm:"index"
        x-nullable: true
      severity:
        type: string
        enum: [info, warning, error, critical]
      category:
        type: string
        enum: ['user', 'metrics']
        x-go-custom-tag: gorm:"default:'user'"
      message:
        type: string
        x-go-custom-tag: gorm:"type:varchar(4096)"
      event_time:
        type: string
        format: date-time
        x-go-custom-tag: gorm:"type:timestamp with time zone"
      request_id:
        type: string
        format: uuid
        description: Unique identifier of the request that caused this event to occur.
      props:
        type: string
        description: Additional properties for the event in JSON format.
        x-go-custom-tag: gorm:"type:text"
...
  error:
    type: object
    required:
      - kind
      - id
      - href
      - code
      - reason
    properties:
      kind:
        type: string
        enum: ['Error']
        description: Indicates the type of this object. Will always be 'Error'.
      id:
        type: integer
        format: int32
        description: Numeric identifier of the error.
        minimum: 400
        maximum: 504
      href:
        type: string
        description: Self link.
      code:
        type: string
        description: Globally unique code of the error, composed of the unique identifier of the API and the numeric identifier of the error. For example, if the numeric identifier of the error is 93 and the identifier of the API is assisted_install then the code will be ASSISTED-INSTALL-93.
      reason:
        type: string
        description: Human-readable description of the error.
...
  infra_error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
        description: Numeric identifier of the error.
        minimum: 401
        maximum: 403
      message:
        type: string
        description: Human-readable description of the error.
...
  operator-property:
    type: object
    properties:
      name:
        type: string
        description: Name of the property
      data_type:
        type: string
        enum: ['boolean', 'string', 'integer', 'float']
        description: Type of the property
      mandatory:
        type: boolean
        description: Indicates whether the property is reqired
      options:
        type: array
        items:
          type: string
        description: Values to select from
      description:
        type: string
        description: Description of a property
      default_value:
        type: string
        description: Default value for the property
  operator-properties:
    type: array
    items:
      $ref: '#/definitions/operator-property'
...